requirements.txt
# Librerias necesarias para el funcionamiento del proyecto
# Instalacion:

# 1. Terminal: "python -m venv env" o "venv env"
# 2. Activan el entorno virtual: env\Scripts\activate
# 3. Terminal: pip install -r requirements.txt

# Dependencias:

FastAPI
uvicorn
python-dotenv
pyjwt
supabase